下载项选择优化，数据库查询关联生命周期

+++ app/src/main/java/com/xyoye/dandanplay/utils/database/builder/DeleteBuilder.java
@@ -52,7 +52,8 @@ public class DeleteBuilder {
             clause = "";
         }
 
-        sqLiteDatabase.delete(DataBaseInfo.getTableNames()[tablePosition], clause, args);
+        if (sqLiteDatabase.isOpen())
+            sqLiteDatabase.delete(DataBaseInfo.getTableNames()[tablePosition], clause, args);
     }
 
     public void postExecute() {
